Potの更新とEndへの遷移

+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/model/Pot.kt
@@ -2,7 +2,7 @@ package com.ebata_shota.holdemstacktracker.domain.model
 
 data class Pot(
     val id: PotId,
-    val potNumber: Long,
+    val potNumber: Int,
     val potSize: Int,
     val involvedPlayerIds: List<PlayerId>,
     val isClosed: Boolean
